+++ protocols/rendezvous/tests/rendezvous.rs
@@ -21,7 +21,7 @@
 #[macro_use]
 pub mod harness;
 
-use crate::harness::{await_events_or_timeout, new_swarm, SwarmExt};
+use crate::harness::{await_event_or_timeout, await_events_or_timeout, new_swarm, SwarmExt};
 use futures::stream::FuturesUnordered;
 use futures::StreamExt;
 use libp2p_core::identity;
@@ -170,13 +170,18 @@ async fn discover_allows_for_dial_by_peer_id() {
     alice
         .behaviour_mut()
         .register(namespace.clone(), roberts_peer_id, None);
-    bob.behaviour_mut()
-        .discover(Some(namespace.clone()), None, None, roberts_peer_id);
-
     assert_behaviour_events! {
         alice: rendezvous::client::Event::Registered { .. },
-        bob: rendezvous::client::Event::Discovered { .. },
         || { }
+    };
+
+    bob.behaviour_mut()
+        .discover(Some(namespace.clone()), None, None, roberts_peer_id);
+    assert_behaviour_events! {
+        bob: rendezvous::client::Event::Discovered { registrations,.. },
+        || {
+            assert!(!registrations.is_empty());
+        }
     };
 
     let alices_peer_id = *alice.local_peer_id();
@@ -275,13 +280,17 @@ async fn registration_on_clients_expire() {
     alice
         .behaviour_mut()
         .register(namespace.clone(), roberts_peer_id, Some(registration_ttl));
-    bob.behaviour_mut()
-        .discover(Some(namespace), None, None, roberts_peer_id);
-
     assert_behaviour_events! {
         alice: rendezvous::client::Event::Registered { .. },
-        bob: rendezvous::client::Event::Discovered { .. },
         || { }
+    };
+    bob.behaviour_mut()
+        .discover(Some(namespace), None, None, roberts_peer_id);
+    assert_behaviour_events! {
+        bob: rendezvous::client::Event::Discovered { registrations,.. },
+        || {
+            assert!(!registrations.is_empty());
+        }
     };
 
     tokio::time::sleep(Duration::from_secs(registration_ttl + 5)).await;