+++ CHANGELOG.md
@@ -42,7 +42,26 @@
 
 # `libp2p` facade crate
 
-## Version 0.40.0 [unreleased]
+## Version 0.41.0 [unreleased]
+
+- Update individual crates.
+    - `libp2p-kad`
+    - `libp2p-websocket`
+- Forward `wasm-bindgen` feature to `futures-timer`, `instant`, `parking_lot`, `getrandom/js` and `rand/wasm-bindgen`.
+
+## Version 0.40.0-rc.3 [2021-10-27]
+
+- Update individual crates.
+    - `libp2p-core`
+    - `libp2p-identify`
+    - `libp2p-swarm`
+
+## Version 0.40.0-rc.2 [2021-10-15]
+
+- Update individual crates.
+    - `libp2p-kad`
+
+## Version 0.40.0-rc.1 [2021-10-15]
 
 - Update individual crates.
     - `libp2p-core`